refactor: drop the 2FA cancellation re-export from the twoFactor barrel
